Use PRI?64 instead of ll? in XID format

+++ src/backend/access/transam/commit_ts.c
@@ -282,8 +282,8 @@ TransactionIdGetCommitTsData(TransactionId xid, TimestampTz *ts,
 	if (!TransactionIdIsValid(xid))
 		ereport(ERROR,
 				(errcode(ERRCODE_INVALID_PARAMETER_VALUE),
-				 errmsg("cannot retrieve commit timestamp for transaction %llu",
-						(unsigned long long) xid)));
+				 errmsg("cannot retrieve commit timestamp for transaction %" PRIu64,
+						xid)));
 	else if (!TransactionIdIsNormal(xid))
 	{
 		/* frozen and bootstrap xids are always committed far in the past */
